restructure .devcontainer folder
