+++ templates/index.ts
@@ -43,13 +43,13 @@ const templates: { name: string, templateFunction: TemplateFunction }[] = [
 	{ name: "_babelrc.ts", templateFunction: require("./_babelrc") },
 	{ name: "_devcontainer/devcontainer.json.ts", templateFunction: require("./_devcontainer/devcontainer.json") },
 	{ name: "_devcontainer/docker-compose.yml.ts", templateFunction: require("./_devcontainer/docker-compose.yml") },
-	{ name: "_devcontainer/nginx.conf.ts", templateFunction: require("./_devcontainer/nginx.conf") },
+	{ name: "_devcontainer/iobroker/_Dockerfile.ts", templateFunction: require("./_devcontainer/iobroker/_Dockerfile") },
+	{ name: "_devcontainer/nginx/nginx.conf.ts", templateFunction: require("./_devcontainer/nginx/nginx.conf") },
 	{ name: "_devcontainer/parcel/run-sync.sh.ts", templateFunction: require("./_devcontainer/parcel/run-sync.sh") },
 	{ name: "_devcontainer/parcel/run.sh.ts", templateFunction: require("./_devcontainer/parcel/run.sh") },
 	{ name: "_devcontainer/parcel/sync.sh.ts", templateFunction: require("./_devcontainer/parcel/sync.sh") },
 	{ name: "_devcontainer/parcel/_Dockerfile.ts", templateFunction: require("./_devcontainer/parcel/_Dockerfile") },
 	{ name: "_devcontainer/README.md.ts", templateFunction: require("./_devcontainer/README.md") },
-	{ name: "_devcontainer/_Dockerfile.ts", templateFunction: require("./_devcontainer/_Dockerfile") },
 	{ name: "_eslintrc_javascript.json.ts", templateFunction: require("./_eslintrc_javascript.json") },
 	{ name: "_eslintrc_typescript.js.ts", templateFunction: require("./_eslintrc_typescript.js") },
 	{ name: "_github/ISSUE_TEMPLATE/bug_report.md.ts", templateFunction: require("./_github/ISSUE_TEMPLATE/bug_report.md") },
